Move CefDownloadItemCallbackWrapper into Internals folder/namespace
