+++ CefSharp.Core/CefSharp.Core.vcxproj
@@ -243,10 +243,10 @@
   <ItemGroup>
     <ClInclude Include="BrowserSettings.h" />
     <ClInclude Include="Cef.h" />
-    <ClInclude Include="CefDownloadItemCallbackWrapper.h" />
     <ClInclude Include="Internals\AutoLock.h" />
     <ClInclude Include="Internals\CefBrowserHostWrapper.h" />
     <ClInclude Include="Internals\CefContextMenuParamsWrapper.h" />
+    <ClInclude Include="Internals\CefDownloadItemCallbackWrapper.h" />
     <ClInclude Include="Internals\CefDragDataWrapper.h" />
     <ClInclude Include="Internals\CefFrameWrapper.h" />
     <ClInclude Include="Internals\CefSharpBrowserWrapper.h" />
