+++ CefSharp.Core/Internals/CefDownloadItemCallbackWrapper.h
